+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.0-rc1 - 2018-12-19 
+
+[Full Changelog](https://github.com/netzmacht/release-notifier/compare/0.3.2...0.4.0)
+
+### Added
+
+- Add ability to define a wait interval between each publish action
+- Add options to the configuration which are passed to the renderer
+- Add twitter publisher
+- Add ability to mark new topics as sticky
+
 ## 0.3.2 - 2018-11-01
 
 [Full Changelog](https://github.com/netzmacht/release-notifier/compare/0.3.1...0.3.2)